+++ sider/transaction.py
@@ -13,6 +13,7 @@
 from __future__ import absolute_import
 import warnings
 import functools
+import traceback
 from redis.client import WatchError
 from .exceptions import DoubleTransactionError, ConflictError
 from .warnings import TransactionWarning
@@ -40,14 +41,18 @@ class Transaction(object):
     def __enter__(self):
         context = self.session.context_locals
         if context['transaction'] is not None:
-            raise DoubleTransactionError('there is already a transaction for '
-                                         'the session ' + repr(self.session))
+            raise DoubleTransactionError(
+                'there is already a transaction for the session ' +
+                repr(self.session) + self.format_enter_stack()
+            )
         context['transaction'] = self
         client = self.session.client
         context['original_client'] = client
         pipeline = client.pipeline()
         self.session.client = pipeline
         pipeline.watch(*self.keys)
+        if self.session.verbose_transaction_error:
+            self.enter_stack = traceback.format_stack()
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
@@ -75,11 +80,78 @@ class Transaction(object):
         """
         if self.commit_phase:
             warnings.warn('the transaction already is on commit phase; '
-                          'begin_commit() method seems called twice or more',
+                          'begin_commit() method seems called twice or more' +
+                          self.format_commit_stack(),
                           category=TransactionWarning, stacklevel=2)
             return
         self.commit_phase = True
+        if self.session.verbose_transaction_error:
+            self.commit_stack = traceback.format_stack()
         self.session.client.multi()
+
+    def format_enter_stack(self, indent=4,
+                           title='Traceback where the transaction entered:'):
+        """Makes :attr:`enter_stack` text readable.
+        If its :attr:`session.verbose_transaction_error
+        <sider.session.Session.verbose_transaction_error>` is not ``True``,
+        it will simply return an empty string.
+
+        :param indent: the number of space character for indentation.
+                       default is 4
+        :type indent: :class:`numbers.Integral`
+        :param title: the title string of the formatted traceback. optional
+        :type title: :class:`basestring`
+        :returns: the formatted :attr:`enter_stack` text
+        :rtype: :class:`basestring`
+
+        .. note::
+
+           It's totally for internal use.
+
+        """
+        if self.session.verbose_transaction_error:
+            try:
+                stack = self.enter_stack
+            except AttributeError:
+                return ''
+            indent_str = ' ' * indent
+            tb = '\n'.join(indent_str + line
+                           for frame in stack
+                           for line in frame.splitlines())
+            return '\n{0}{1}\n{2}'.format(indent_str, title, tb)
+        return ''
+
+    def format_commit_stack(self, indent=4,
+                            title='Traceback of previous begin_commit() call:'):
+        """Makes :attr:`commit_stack` text readable.
+        If its :attr:`session.verbose_transaction_error
+        <sider.session.Session.verbose_transaction_error>` is not ``True``,
+        it will simply return an empty string.
+
+        :param indent: the number of space character for indentation.
+                       default is 4
+        :type indent: :class:`numbers.Integral`
+        :param title: the title string of the formatted traceback. optional
+        :type title: :class:`basestring`
+        :returns: the formatted :attr:`commit_stack` text
+        :rtype: :class:`basestring`
+
+        .. note::
+
+           It's totally for internal use.
+
+        """
+        if self.session.verbose_transaction_error:
+            try:
+                stack = self.commit_stack
+            except AttributeError:
+                return ''
+            indent_str = ' ' * indent
+            tb = '\n'.join(indent_str + line
+                           for frame in stack
+                           for line in frame.splitlines())
+            return '\n{0}{1}\n{2}'.format(indent_str, title, tb)
+        return ''
 
 
 def manipulative(function):